Add files via upload
Script is still in the debugging phase, It's guaranteed that it'll crash, but you can get an idea of what I'm doing from looking at the code.
Some parts have been butchered out from JhonnyOThan's TwitchPlaysKSP's execute_node to follow up on the idea he had. This file could be considered a fork from that, maybe.